feat: 비로그인 시 localstorage를 이용해 장바구니 저장

+++ src/apis/baskets.js
@@ -1,9 +1,12 @@
 import request from './utils/request';
 import apis from './index';
+import auth from '../utils/auth';
+import localstorage, { STORAGE_KEYS } from '../utils/localstorage';
 
 const baskets = {
   async getAll() {
-    const baskets = await request.get('/baskets');
+    const isGuest = !auth.isLoggedIn();
+    const baskets = isGuest ? localstorage.get(STORAGE_KEYS.BASKETS) || [] : await request.get('/baskets');
     const getProductPromise = baskets.map(({ product_id }) => apis.products.get(product_id));
     const products = await Promise.all(getProductPromise);
 
@@ -14,12 +17,31 @@ const baskets = {
     }));
   },
 
-  update(id, amount) {
-    return request.post(`/baskets/${id}`, { amount });
+  async update(id, amount) {
+    if (auth.isLoggedIn()) {
+      return request.post(`/baskets/${id}`, { amount });
+    }
+    const baskets = localstorage.get(STORAGE_KEYS.BASKETS) || [];
+    const targetIndex = baskets.findIndex((basket) => basket.id === id);
+    const changedBasket = {
+      ...baskets[targetIndex],
+      amount,
+    };
+    const changed = [
+      ...baskets.slice(0, targetIndex),
+      changedBasket,
+      ...baskets.slice(targetIndex + 1, baskets.length),
+    ];
+    localstorage.set(STORAGE_KEYS.BASKETS, changed);
   },
 
-  remove(id) {
-    return request.delete(`/baskets/${id}`);
+  async remove(id) {
+    if (auth.isLoggedIn()) {
+      return request.delete(`/baskets/${id}`);
+    }
+    const baskets = localstorage.get(STORAGE_KEYS.BASKETS) || [];
+    const targetIndex = baskets.findIndex((basket) => basket.id === id);
+    localstorage.set(STORAGE_KEYS.BASKETS, baskets.splice(targetIndex, 1));
   },
 };
